코드 파편화 수정, 변수 삭제, 세션, 쿠키 관리 컨트롤러에서 수행

+++ src/main/java/toy/recipit/common/Constants.java
@@ -64,7 +64,7 @@ public final class Constants {
 
     public interface UserLogin {
         int LOGIN_FAIL_COUNT_INITIAL = 0;
-        long AUTO_LOGIN_EXPIRATION_DAYS = 7L;
+        int AUTO_LOGIN_EXPIRATION_DAYS = 7;
         String AUTO_LOGIN_COOKIE_NAME = "AUTO_LOGIN_TOKEN";
     }
 }
